// src/types/partials/CodeSystemFilter.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';

export type TCodeSystemFilter = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    code: String;
    description?: String;
    operator: String[];
    value: String;
};